requirements.txt
# Framework Web
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Base de Datos - MySQL
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# Base de Datos - MongoDB
pymongo==4.6.1

# Base de Datos - Redis
redis==5.0.1

# Base de Datos - Neo4j
neo4j==5.15.0

# Autenticación
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
bcrypt==4.1.2

# Utilidades
python-dotenv==1.0.0
marshmallow==3.20.1

# Validación
email-validator==2.1.0

# Desarrollo
flask-cors==4.0.0